requirements.txt
pytest
selenium
openpyxl
PyYAML
pytest-html

pip install selenium pytest pytest-html openpyxl webdriver-manager
